Zombie (it would be cool if they used zombie figurines thou)
So, guys, I found this on GitHub accidentally while browsing what trained model we should use, and this guy was also doing cards! After looking at this code, I am 99.99% sure either this was posted by PiWars or someone doing PiWars because chances are low that someone not in PiWars is trying to do the same thing as us just for fun. Admittedly, this could be better, but it is terrific. It also uses PID; is that a problem?