Update zombiev2
So, I added some things to make this code more practical. However, because our robot needs two inputs for each motor, even though one is not required, I have changed my code to fit our needs. The hard bit of the code is correct, but whether my bit is correct is another question. Might I ask someone to check it to see if it is correct?