Update lava.py
I decided to not use threading and asyncio since I don't know much about threading. I decided to just spam checking kill switch everywhere. I mean it is not like we need robot to stop the exact millisecond to the exact decimal point the order was given out. Maximum of 0.2 seconds delay is pretty good for us, since we don't need that much.