Update main.py
finished test1() function for 5th of march session. Added Ethan's motor update function to the move() and spin() functions so that the robot's position is updated each time we move it. 

Re-ordered code, putting motor code after all of the vector computation. 

Still need to find the value for the motor turn rate.